Move web assets to dist and update build
