+++ capacitor.config.ts
@@ -3,7 +3,7 @@ import { CapacitorConfig } from '@capacitor/cli';
 const config: CapacitorConfig = {
   appId: 'auditory.trainer',
   appName: 'Auditory Trainer',
-  webDir: '.',
+  webDir: 'dist',
   bundledWebRuntime: false,
 };
